Moving coral-framework tests over and updating mocha to transpile them.
